Speed up restart_band_structure.py, fix gllb-io for parallel, fix restart_band_structure for parallel. Some information is still lost on io since no matter how tight the convergence criteria is, the eigenvalue difference is still ~0.0001eV. Maybe some finegrid restriction releated?
git-svn-id: https://svn.fysik.dtu.dk/projects/gpaw/trunk@4157 6419b465-d902-0410-997c-be2266baf723

+++ test/restart_band_structure.py
@@ -4,39 +4,32 @@ from ase.calculators import numeric_force
 from gpaw.utilities import equal
 import os
 from gpaw import setup_paths
+from gpaw.mpi import world
 setup_paths.insert(0, '.')
 
-for xc in ['GLLBSC','LDA']:
-    a = 5.404
-    bulk = Atoms(symbols='Si8',
-                 positions=[(0, 0, 0),
-                            (0, 0.5, 0.5),
-                            (0.5, 0, 0.5),
-                            (0.5, 0.5, 0),
-                            (0.25, 0.25, 0.25),
-                            (0.25, 0.75, 0.75),
-                            (0.75, 0.25, 0.75),
-                            (0.75, 0.75, 0.25)],
-                 pbc=True)
-    bulk.set_cell((a, a, a), scale_atoms=True)
-    n = 20
-    calc = GPAW(gpts=(n, n, n),
-                nbands=24,
+for xc in ['LDA','GLLBSC']:
+    a = 4.23
+    bulk = Atoms('Si2', cell=(a, a, a), pbc=True,
+              scaled_positions=[[0, 0, 0], [.5, .5, .5]])
+    calc = GPAW(h=0.25,
+                nbands=8,
                 width=0.01,
-                kpts=(3, 3, 3), convergence={'eigenvalue':1e-12}, xc=xc)
+                kpts=(3, 3, 3), convergence={'eigenstates':1e-12, 'bands':8}, xc=xc, eigensolver='cg')
 
     bulk.set_calculator(calc)
     bulk.get_potential_energy()
     print calc.get_ibz_k_points()
-    old_eigs = calc.get_eigenvalues(kpt=1)
+    old_eigs = calc.get_eigenvalues(kpt=3)
     calc.write('Si_gs.gpw')
     del bulk
     del calc
-    bulk, calc = restart('Si_gs.gpw', fixdensity=True, kpts=[[0,0,0],[1./3,0,0]])
+    bulk, calc = restart('Si_gs.gpw', fixdensity=True, kpts=[[0,0,0],[1./3,1./3,1./3]])
     bulk.get_potential_energy()
 
-    os.remove('Si_gs.gpw')
-    diff = calc.get_eigenvalues(kpt=1)[1:16]-old_eigs[1:16]
-    print "occ. eig. diff.", diff
-    error = max(abs(diff))
-    assert error < 2e-5
+    if world.rank == 0:
+        os.remove('Si_gs.gpw')
+    diff = calc.get_eigenvalues(kpt=1)[:6]-old_eigs[:6]
+    if world.rank == 0:
+        print "occ. eig. diff.", diff
+        error = max(abs(diff))
+        assert error < 5e-6
